+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 2.0
 =====
+
 * Overhaul of the API-s
 * HTML::breadcrumbs can now accept undefined number of crumbs in one call, using func_get_args
 * Date::mysql_date accepts DateTime and throws InvalidArgumentException
@@ -12,6 +13,7 @@
 
 1.4
 =====
+
 * New class: Valid for extra validation rules
 * Deprecated LESS support from Assets class
 * Assets now have presets (Asset::preset())